requirements.txt
crewai>=0.28.0
crewai-tools>=0.1.6
google-generativeai>=0.3.2
langchain>=0.1.20
langchain-core>=0.1.52
langchain-google-genai>=1.0.1
langchain-openai>=0.1.7
langchain-community>=0.0.38
langchain-pinecone>=0.1.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
pinecone-client>=3.2.0
loguru>=0.7.2
pypdf>=3.17.0
langchain-text-splitters>=0.0.1
tiktoken>=0.5.2
sentence-transformers>=2.2.2